Now up-to-date with the progress I made in my own SVN repository. Lots of changes to the audio output interface - it is now a lot cleaner.
git-svn-id: svn://scm.orgis.org/mpg123/trunk@159 35dc7657-300d-0410-a2e5-dc2837fedb53

+++ src/audio_macosx.c
@@ -223,7 +223,7 @@ void start(int n)
 
 int audio_open(struct audio_info_struct *ai)
 {
-	long size;
+	UInt32 size;
 	AudioStreamBasicDescription format;
 	#if MOSX_USES_SEM
 	char s[10];
